+++ Tavisca.Tripster.Contracts/Error/ErrorCode.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Tavisca.Tripster.Contracts
+{
+    public class ErrorCode
+    {
+        public const int InternalServerError = 500;
+    }
+}
