+++ Tavisca.Tripster.Contracts/Error/ErrorMessages.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Tavisca.Tripster.Contracts
+{
+    public class ErrorMessages
+    {
+        public const string InternalServerError = "Internal Server Error";
+    }
+}
