--- Tavisca.Tripster.Contracts/Service/TransferObject.cs
@@ -1,8 +1,0 @@
-﻿namespace Tavisca.Tripster.Core.Validation
-{
-    public class TransferObject<TEntity> where TEntity : class
-    {
-        public string ErrorMessage { get; set; } = null;
-        public TEntity ModelObject { get; set; } = null;
-    }
-}
