+++ Tavisca.Tripster.Contracts/Tavisca.Tripster.Contracts.csproj
@@ -1,15 +1,13 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp2.1</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="FluentAssertions" Version="5.9.0" />
+    <PackageReference Include="FluentValidation" Version="8.5.1" />
     <PackageReference Include="MongoDB.Driver" Version="2.9.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Tavisca.Tripster.Data\Tavisca.Tripster.Data.csproj" />
   </ItemGroup>
 
 </Project>
