+++ Tavisca.Tripster.Contracts/interface/IRepository.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+
+namespace Tavisca.Tripster.Contracts { 
+    public interface IRepository<TEntity> where TEntity : class
+    {
+        TEntity Get(Guid id);
+        IEnumerable<TEntity> GetAll();
+        void Add(TEntity entity);
+        void Delete(Guid id);
+        void Update(Guid id, TEntity entity);
+        
+    }
+
+}
