+++ Tavisca.Tripster.Contracts/interface/TransferObject.cs
@@ -1,0 +1,10 @@
+﻿
+
+namespace Tavisca.Tripster.Contracts
+{
+    public class TransferObject<TEntity> where TEntity : class
+    {
+        public string ErrorMessage { get; set; } = null;
+        public TEntity ModelObject { get; set; } = null;
+    }
+}
