--- Tavisca.Tripster.MongoDB/Tavisca.Tripster.MongoDB.csproj
@@ -1,11 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Tavisca.Tripster.Contracts\Tavisca.Tripster.Contracts.csproj" />
-  </ItemGroup>
-
-</Project>
