+++ Tavisca.Tripster.Web/Tavisca.Tripster.Web.csproj
@@ -1,11 +1,13 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp2.1</TargetFramework>
     <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="FluentAssertions" Version="5.9.0" />
+    <PackageReference Include="FluentValidation" Version="8.5.1" />
     <PackageReference Include="Microsoft.AspNetCore.App" />
     <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />
     <PackageReference Include="MongoDB.Driver" Version="2.9.2" />
@@ -15,7 +17,7 @@
   <ItemGroup>
     <ProjectReference Include="..\Tavisca.Tripster.Contracts\Tavisca.Tripster.Contracts.csproj" />
     <ProjectReference Include="..\Tavisca.Tripster.Core\Tavisca.Tripster.Core.csproj" />
-    <ProjectReference Include="..\Tavisca.Tripster.MongoDB\Tavisca.Tripster.MongoDB.csproj" />
+    <ProjectReference Include="..\Tavisca.Tripster.MongoDB\Tavisca.Tripster.Dal.csproj" />
   </ItemGroup>
 
 </Project>
