+++ Tavisca.Tripster.sln
@@ -7,13 +7,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tavisca.Tripster.Web", "Tav
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tavisca.Tripster.Contracts", "Tavisca.Tripster.Contracts\Tavisca.Tripster.Contracts.csproj", "{1C45AA41-325E-4516-9206-1E6C4E75A509}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tavisca.Tripster.Data", "Tavisca.Tripster.Data\Tavisca.Tripster.Data.csproj", "{F3F856B9-B366-43E4-9DD0-4D37AA0410CC}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tavisca.Tripster.Core", "Tavisca.Tripster.Core\Tavisca.Tripster.Core.csproj", "{1F607D51-5F20-4A71-85E5-74A0E52BDFF1}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tavisca.Tripster.Tests", "Tavisca.Tripster.Tests\Tavisca.Tripster.Tests.csproj", "{2EEA449D-0851-41D2-89FF-50302F41E047}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tavisca.Tripster.MongoDB", "Tavisca.Tripster.MongoDB\Tavisca.Tripster.MongoDB.csproj", "{11A99D08-0D3D-4D9C-9858-62B32622E46C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tavisca.Tripster.Dal", "Tavisca.Tripster.MongoDB\Tavisca.Tripster.Dal.csproj", "{11A99D08-0D3D-4D9C-9858-62B32622E46C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -29,10 +27,6 @@ Global
 		{1C45AA41-325E-4516-9206-1E6C4E75A509}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1C45AA41-325E-4516-9206-1E6C4E75A509}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1C45AA41-325E-4516-9206-1E6C4E75A509}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F3F856B9-B366-43E4-9DD0-4D37AA0410CC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F3F856B9-B366-43E4-9DD0-4D37AA0410CC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F3F856B9-B366-43E4-9DD0-4D37AA0410CC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F3F856B9-B366-43E4-9DD0-4D37AA0410CC}.Release|Any CPU.Build.0 = Release|Any CPU
 		{1F607D51-5F20-4A71-85E5-74A0E52BDFF1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{1F607D51-5F20-4A71-85E5-74A0E52BDFF1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1F607D51-5F20-4A71-85E5-74A0E52BDFF1}.Release|Any CPU.ActiveCfg = Release|Any CPU
